\name{compare}
\alias{compare}
\title{Comparison Of GWAS studies.}
\description{Compares GWAS studies.}
\usage{compare(script_file, pval_threshold, maf_threshold)}
\arguments{
	\item{script_file}{The name of the METAL script file.}
	\item{pval_threshold}{A threshold for the P-value.}
	\item{maf_threshold}{A threshold for the MAF.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}