\name{dispersion_check}
\alias{dispersion_check}
\title{Dispersion Check.}
\description{Dispersion check.}
\usage{dispersion_check(script, sample_sizes = NULL, plot = TRUE)}
\arguments{
	\item{script}{
		Name of a textual input file with processing instructions. 
		The file should contain the names and locations of all GWAS data files to be processed 
		along with basic information from each individual study, and instructions for the quality check.
	}
	\item{sample_sizes}{Optional vector of sample sizes.}
	\item{plot}{Produce plot.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}