\name{gene2snpByName}
\alias{gene2snpByName}
\title{Gene to SNP annotation.}
\description{Annotates genes to SNPs.}
\usage{gene2snpByName(names, genes, snps, output, append, multicolumn)}
\arguments{
	\item{names}{No description.}
	\item{genes}{The name of an input file with genes.}
	\item{snps}{The name of an input file with SNPs.}
	\item{output}{The name of an output file.}
	\item{append}{If true, then writes all columns from an input file into a result file.}
	\item{multicolumn}{No description.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}