\name{independent_snps}
\alias{independent_snps}
\title{Selection of independent SNPs.}
\description{Function selects independent SNPs based on the genome-wide association P-values and linkage disequilibrium.}
\usage{independent_snps(script)}
\arguments{
	\item{script}{
		Name of a textual input file with processing instructions. 
		The file should contain the names and locations of all input data files to be processed 
		along with their basic descriptions.
	}
}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}