\name{kusk_check}
\alias{kusk_check}
\title{Comparison Of Skewness and Kurtosis of Effect between GWAS studies.}
\description{Plots Skewness vs Kurtosis of Effect in GWAS studies.}
\usage{kusk_check(summary_file, worst, plot)}
\arguments{
	\item{summary_file}{A name of the summary file.}
	\item{worst}{Percentage of worst p-values.}
	\item{plot}{Produce plot.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}