\name{pgwasqc}
\alias{pgwasqc}
\title{Quality Control Of GWA Data}
\description{Performs the quality control of data from Genome-Wide Association Studies (GWAS).}
\usage{pgwasqc(script, cpu)}
\arguments{
	\item{script}{
		Name of a textual input file with processing instructions. 
		The file should contain the names and locations of all GWAS data files to be processed 
		along with basic information from each individual study, and instructions for the quality check.
	}
	\item{cpu}{
		Number of CPUs.
	}
}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{plot}
\keyword{package}
\examples{

}
