\name{snp2geneHS}
\alias{snp2geneHS}
\title{SNP to gene annotation.}
\description{Annotates SNPs to genes assuming hotspots.}
\usage{snp2geneHS(chromosome, snps, ld, hotspots, genes, prefix)}
\arguments{
	\item{chromosome}{}
	\item{snps}{}
	\item{ld}{}
	\item{hotspots}{}
	\item{genes}{}
	\item{prefix}{}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}