\name{annotate}
\alias{annotate}
\title{SNPs annotation with genes.}
\description{For provided SNPs finds all genes in specified deviation (e.g. 0, +/- 50000, +/- 100000, ... ).}
\usage{annotate(script)}
\arguments{
	\item{script}{
		Name of a textual input file with processing instructions. 
		The file should contain the names and locations of all GWAS data files to be annotated 
		along with basic information from each individual study.
	}
}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}