\name{dispersion_check}
\alias{dispersion_check}
\title{Dispersion Check.}
\description{Dispersion check.}
\usage{dispersion_check(summary_file, sample_sizes, plot)}
\arguments{
	\item{summary_file}{A name of the summary file.}
	\item{sample_sizes}{Optional vector of sample sizes.}
	\item{plot}{Produce plot.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}