\name{snp2geneHS}
\alias{snp2geneHS}
\title{SNP to gene annotation.}
\description{Annotates SNPs to genes assuming hotspots.}
\usage{snp2geneHS(chromosome, snps, ld, hotspots, genes, prefix)}
\arguments{
	\item{chromosome}{No description.}
	\item{snps}{No description.}
	\item{ld}{No description.}
	\item{hotspots}{No description.}
	\item{genes}{No description.}
	\item{prefix}{No description.}
}
\note{Under development.}
\author{Daniel Taliun, Christian Fuchsberger, Cristian Pattaro}
\keyword{misc}
\keyword{utilities}
\keyword{package}